+++ ControlRoom/Simulator UI/ControlScreens/AppView.swift
@@ -85,11 +85,6 @@ struct AppView: View {
                     Button("Revoke", action: revokePrivacy)
                     Button("Reset", action: resetPrivacy)
                 }
-
-                HStack {
-                    TextField("URL to open", text: $preferences.lastOpenURL)
-                    Button("Open URL", action: openURL)
-                }
             }
 			.disabled(!isApplicationSelected)
 
@@ -111,7 +106,7 @@ struct AppView: View {
 
         }
         .tabItem {
-            Text("App")
+            Text("Applications")
         }
         .padding()
         .alert(isPresented: $shouldShowUninstallConfirmationAlert) {
@@ -151,11 +146,6 @@ struct AppView: View {
     /// Removes the identified app from the device.
     func uninstallApp() {
         SimCtl.uninstall(simulator.udid, appID: selectedApplication.bundleIdentifier)
-    }
-
-    /// Opens a URL in the appropriate device app.
-    func openURL() {
-        SimCtl.openURL(simulator.udid, URL: preferences.lastOpenURL)
     }
 
     /// Grants some type of permission to the app.
